refactor: move t.String to PrimitiveType
